Upgrade all package (Update dependencies) with issue here # https://github.com/Algram/ytdl-webserver/issues/8
Fix security vulnerabilities with the way is update `Hapi` framework and `webpack-dev-server` to latest version.
Remove `babel-preset-es2015` and using `@babel/preset-env` for instead.
Updated `webpack, @babel/preset-react, @babel/preset-env, @babel/core, babel-loader` to latest version because it depend to some security vulnerabilities
Add `webpack-cli` because webpack 4 require it.
Add `mini-css-extract-plugin` instead of `extract-text-webpack-plugin` (remove it) because webpack 4 do not support it.

Updated react and all dependencies of it
Updated the rest dependencies

Add arguments --mode development and --mode production to `dev` and `build` in scripts for better performance.